Specify attributes for Project alignment
A project should have an attribute indicating whether the UI will align the display of the project in the center of the page, or split down the center of the page with content on the left and right. Also, a project should have an attribute indicating how many elements of content to put on the left side of a split display (implying that the remainder of elements will be placed on the right side).

This will enable more variety in the UI layout by enabling some projects to be shown with visual emphasis on the left or right side instead of just the center.