Refactor Project model into generic Section model
Previously, the Project model was a way of distinguishing different projects on the projects page. Viewed in a more abstract way, these divisions of a page could be called 'sections'. This approach is more generalized and would still allow the projects page to have multiple sections, while enabling other pages to have multiple sections as well without needing to make any new models.

This commit replaces the Project model with a new Section model.