Remove copy attribute from Paragraph
This commit removes the copy attribute from the Paragraph model since it is no longer being used in this context and is left over from an earlier schema. Copy is an attribute of the SentenceGroup model, and a SentenceGroup instance belongs to a Paragraph instance, but the Paragraph model itself doesn't need a copy attribute.