Specify Pages in the API response
A GET request to '/api/v1/pages.json' should result in a response where the body is an array of objects that each represent a page. Each page object should contain a title and an array of objects that each represent a section. Each section object can contain objects representing images and paragraphs, and a paragraph object can contain objects representing sentence groups and links.